Fixed event group handling
Cleaned event bits

created switched logic behind reding the event bits. The switching between calculating task with start & stopping tasks and reset the calcualtions is working.

Now both tasks are first waiting until the get selected. Then they are checking if they have to reset the value or if they can start calculating.

The UI task is now showing selection and when start is pressed it shows the calculation